requirements.txt
fastapi 
pydantic 
pymongo 
openai 
python-dotenv 
reportlab 
motor